+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Pool.hs
@@ -260,7 +260,8 @@ poolDelegationTransition = do
           tellEvent $ RegisterPool sppId
           pure $
             ps
-              & psStakePoolsL %~ Map.insert sppId (mkStakePoolState (pp ^. ppPoolDepositCompactL) stakePoolParams)
+              & psStakePoolsL
+                %~ Map.insert sppId (mkStakePoolState (pp ^. ppPoolDepositCompactL) mempty stakePoolParams)
               & psVRFKeyHashesL %~ updateVRFKeyHash
         -- re-register Pool
         Just stakePoolState -> do
@@ -276,7 +277,7 @@ poolDelegationTransition = do
                       Nothing -> Map.insert sppVrf (knownNonZeroBounded @1)
                       Just futureStakePoolState
                         | futureStakePoolState ^. spsVrfL /= sppVrf ->
-                            (Map.insert sppVrf (knownNonZeroBounded @1))
+                            Map.insert sppVrf (knownNonZeroBounded @1)
                               . Map.delete (futureStakePoolState ^. spsVrfL)
                         | otherwise -> id
                 | otherwise = id
@@ -293,10 +294,17 @@ poolDelegationTransition = do
           -- has been removed from the registered pools).  does it need to pay a
           -- new deposit (at the current deposit amount). But of course, if that
           -- has happened, we cannot be in this branch of the case statement.
+          let futureStakePoolState =
+                mkStakePoolState
+                  (stakePoolState ^. spsDepositL)
+                  -- delegators are set in PoolReap,
+                  -- in order to capture delegations that happened after re-registration but before the end of the epoch
+                  mempty
+                  stakePoolParams
           pure $
             ps
               & psFutureStakePoolsL
-                %~ Map.insert sppId (mkStakePoolState (stakePoolState ^. spsDepositL) stakePoolParams)
+                %~ Map.insert sppId futureStakePoolState
               & psRetiringL %~ Map.delete sppId
               & psVRFKeyHashesL %~ updateFutureVRFKeyHash
     RetirePool sppId e -> do