+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/PoolReap.hs
@@ -152,7 +152,11 @@ poolReapTransition = do
     -- activate future stakePools
     ps =
       ps0
-        { psStakePools = Map.union (ps0 ^. psFutureStakePoolsL) (ps0 ^. psStakePoolsL)
+        { psStakePools =
+            Map.unionWith
+              (\newPoolState oldPoolState -> newPoolState {spsDelegators = spsDelegators oldPoolState})
+              (ps0 ^. psFutureStakePoolsL)
+              (ps0 ^. psStakePoolsL)
         , psFutureStakePools = Map.empty
         }
     cs = cs0 & certPStateL .~ ps
@@ -208,11 +212,12 @@ poolReapTransition = do
       a {casTreasury = casTreasury a <+> fromCompact unclaimed}
       ( cs
           & certDStateL . accountsL
-            %~ removeStakePoolDelegations retired . addToBalanceAccounts refunds
+            %~ removeStakePoolDelegations (delegsToClear cs retired)
+              . addToBalanceAccounts refunds
           & certPStateL . psStakePoolsL %~ (`Map.withoutKeys` retired)
           & certPStateL . psRetiringL %~ (`Map.withoutKeys` retired)
           & certPStateL . psVRFKeyHashesL
-            %~ ( removeVRFKeyHashOccurrences (retiredVRFKeyHashes)
+            %~ ( removeVRFKeyHashOccurrences retiredVRFKeyHashes
                    . (`Map.withoutKeys` danglingVRFKeyHashes)
                )
       )
@@ -225,6 +230,9 @@ poolReapTransition = do
     removeVRFKeyHashOccurrence =
       -- Removes the key from the map if the value drops to 0
       Map.update (mapNonZero (\n -> n - 1))
+    delegsToClear cState pools =
+      foldMap spsDelegators $
+        Map.restrictKeys (cState ^. certPStateL . psStakePoolsL) pools
 
 renderPoolReapViolation ::
   ( EraGov era