+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Pool.hs
@@ -52,7 +52,7 @@ pStateSpec univ = constrained $ \ps ->
     , witness univ (dom_ retiring)
     , assertExplain (pure "dom of retiring is a subset of dom of stakePoolParams") $
         dom_ retiring `subset_` dom_ stakePools
-    , forAll' (rng_ stakePools) $ \_ _ _ _ _ _ _ _ [var|d|] ->
+    , forAll' (rng_ stakePools) $ \_ _ _ _ _ _ _ _ [var|d|] _ ->
         assertExplain (pure "all deposits are greater then (Coin 0)") $ d >=. lit 0
     , assertExplain (pure "dom of stakePoolParams is disjoint from futureStakePoolParams") $
         dom_ stakePools `disjoint_` dom_ futureStakePools